unit-test: refine related unit tests
We need to refine related unit tests.

Fixes: #2511

(cherry picked from commit 00da1270be772e7fd2f63bea9c84a8596672a81e)
